pyproject.toml
[tool.poetry]
name = "libmogra"
version = "0.3.4"
description = "\"A Python Toolbox for Indian (Classical) Music\""
authors = ["Neeraja Abhyankar <neeraja.abhyankar@gmail.com>"]
license = "MIT License"
readme = "README.md"
homepage = "https://github.com/neerajaabhyankar/libmogra"
repository = "https://github.com/neerajaabhyankar/libmogra"
keywords=["music", "indian classical music", "raga", "raag", "taal", "swar", "indian music", "classical music", "hindustani classical music"]
classifiers=[
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
include = ["LICENSE"]

[tool.poetry.scripts]
mogra = "libmogra.raagfinder.main:main"

[tool.poetry.dependencies]
python = "^3.11"
dill = "^0.3.9"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
rich = "^13.9.4"
pytest = "^8.3.4"
argparse = "^1.4.0"
rapidfuzz = "^3.11.0"
plotly = "^5.24.1"
pillow = "^11.0.0"
kaleido = "1.0.0rc11"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
